Move common utility functions to string.ts and MathItem.ts
